Revert "Merge branch 'feature/rendertexture_optimization'"
This reverts commit 75e121c9d67375ea2961f3461aed04c2affed2d9, reversing
changes made to 4d7d87bbe09ec31be2b641d4651af7eb9c7d6397.

+++ src/SFML/Audio/SoundFileReaderWav.cpp
@@ -1,7 +1,7 @@
 ////////////////////////////////////////////////////////////
 //
 // SFML - Simple and Fast Multimedia Library
-// Copyright (C) 2007-2018 Laurent Gomila (laurent@sfml-dev.org)
+// Copyright (C) 2007-2017 Laurent Gomila (laurent@sfml-dev.org)
 //
 // This software is provided 'as-is', without any express or implied warranty.
 // In no event will the authors be held liable for any damages arising from the use of this software.
@@ -155,7 +155,7 @@ Uint64 SoundFileReaderWav::read(Int16* samples, Uint64 maxCount)
     assert(m_stream);
 
     Uint64 count = 0;
-    while ((count < maxCount) && (static_cast<Uint64>(m_stream->tell()) < m_dataEnd))
+    while ((count < maxCount) && (m_stream->tell() < m_dataEnd))
     {
         switch (m_bytesPerSample)
         {
